requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2.14
django-crispy-forms==1.9.2
django-filter==2.3.0
django-nested-admin==3.3.2
idna==2.10
Markdown==3.2.2
psycopg2==2.8.5
python-monkey-business==1.0.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
